add startday fix
