+++ GomezMorinFrontEnd/src/components/InputForm.jsx
@@ -28,21 +28,11 @@ const InputForm = ({
   minLength,
   maxLength,
   min,
-  max
+  max,
+  onChange
 }) => {
   const { register } = useFormContext();
   const [errorMessage, setErrorMessage] = useState("");
-
-  const handleFileChange = (event) => {
-    const file = event.target.files[0];
-    const allowedExtensions = /(\.pdf)$/i;
-
-    if (!allowedExtensions.test(file.name)) {
-      setErrorMessage("Solo se permiten archivos PDF.");
-    } else {
-      setErrorMessage("");
-    }
-  };
 
   return (
     <div className="flex flex-col gap-2 w-full h-full">
@@ -61,11 +51,13 @@ const InputForm = ({
         accept={accept}
         minLength={minLength}
         maxLength={maxLength}
-        min = {min}
-        max = {max}
-        onChange={handleFileChange}
+        min={min}
+        max={max}
+        onChange={onChange}
       />
-      {errorMessage && <p className="font-semibold text-red-400 text-sm">{errorMessage}</p>}
+      {errorMessage && (
+        <p className="font-semibold text-red-400 text-sm">{errorMessage}</p>
+      )}
     </div>
   );
 };
